requirements.txt
discord.py
matplotlib
pytz
